Removing Meta tests and creating a file for them.
